remove paneTransitionScope from ListMemberPickerSheet
